Change peerDependencies to optionalDependencies
optionalDependencies dont spill a bunch of warnings on the screen. I use ioredis and it's reminding me to always install redis and redis-mock.

https://yarnpkg.com/lang/en/docs/dependency-types/

optionalDependencies
Optional dependencies are just that: optional. If they fail to install, Yarn will still say the install process was successful.

This is useful for dependencies that won’t necessarily work on every machine and you have a fallback plan in case they are not installed (e.g. Watchman).