Use redis.expire on touch
This restores the V3 behavior of using `EXPIRE` instead of setting the whole session. This no longer updates the `cookie` object when a `touch` occurs which is more in line with other stores. Note, the main `express-session` project ignores the `cookie` object when hashing the session for changes.

PR: https://github.com/tj/connect-redis/pull/285